Fix for issue #110
Every time, a z-wave device sends a "I'm ready"-Event, the local map entry of this node is re-written.
Thats because a too late abort statement, which should check, if this node is already initialized. As an effect, the Device object instance is getting lost and the `value changes`-Event can't be handled anymore.